+++ client/components/pages/editor/EditorSection.jsx
