+++ client/components/pages/site/themes.js
@@ -1,0 +1,5 @@
+import ClassicTheme from 'themes/Classic/Index'
+
+export {
+  ClassicTheme as Classic
+}
