--- client/components/templates/themes.js
@@ -1,5 +1,0 @@
-import ClassicTheme from './Classic/Index'
-
-export {
-  ClassicTheme as Classic
-}
