+++ client/components/themes/Classic/Footer.jsx
