+++ client/components/themes/Classic/Hero.jsx
